+++ crates/rome_js_analyze/tests/specs/nursery/useIsArray/valid.js
@@ -1,0 +1,4 @@
+Array.isArray(arr)
+arr instanceof Object
+a instanceof ns.Array
+a.x[2] instanceof foo()