+++ crates/rome_service/src/configuration/parse/json/rules.rs
@@ -1672,6 +1672,7 @@ impl VisitNode<JsonLanguage> for Nursery {
                 "useGroupedTypeImport",
                 "useHeadingContent",
                 "useHookAtTopLevel",
+                "useIsArray",
                 "useIsNan",
                 "useLiteralEnumMembers",
                 "useLiteralKeys",
@@ -2332,6 +2333,29 @@ impl VisitNode<JsonLanguage> for Nursery {
                         diagnostics,
                     )?;
                     self.use_hook_at_top_level = Some(rule_configuration);
+                }
+                _ => {
+                    diagnostics.push(DeserializationDiagnostic::new_incorrect_type(
+                        "object or string",
+                        value.range(),
+                    ));
+                }
+            },
+            "useIsArray" => match value {
+                AnyJsonValue::JsonStringValue(_) => {
+                    let mut configuration = RuleConfiguration::default();
+                    self.map_to_known_string(&value, name_text, &mut configuration, diagnostics)?;
+                    self.use_is_array = Some(configuration);
+                }
+                AnyJsonValue::JsonObjectValue(_) => {
+                    let mut rule_configuration = RuleConfiguration::default();
+                    rule_configuration.map_rule_configuration(
+                        &value,
+                        name_text,
+                        "useIsArray",
+                        diagnostics,
+                    )?;
+                    self.use_is_array = Some(rule_configuration);
                 }
                 _ => {
                     diagnostics.push(DeserializationDiagnostic::new_incorrect_type(