+++ xtask/codegen/src/generate_bindings.rs
@@ -39,7 +39,7 @@ pub(crate) fn generate_workspace_bindings(mode: Mode) -> Result<()> {
                         Some(AnyJsParameter::AnyJsFormalParameter(
                             AnyJsFormalParameter::JsFormalParameter(
                                 make::js_formal_parameter(
-                                    make::js_decorator_list(std::iter::empty()),
+                                    make::js_decorator_list([]),
                                     AnyJsBindingPattern::AnyJsBinding(
                                         AnyJsBinding::JsIdentifierBinding(
                                             make::js_identifier_binding(make::ident("params")),
@@ -85,7 +85,7 @@ pub(crate) fn generate_workspace_bindings(mode: Mode) -> Result<()> {
                     make::js_parameter_list(
                         Some(AnyJsParameter::AnyJsFormalParameter(
                             AnyJsFormalParameter::JsFormalParameter(
-                                make::js_formal_parameter(make::js_decorator_list(std::iter::empty()),AnyJsBindingPattern::AnyJsBinding(
+                                make::js_formal_parameter(make::js_decorator_list([]),AnyJsBindingPattern::AnyJsBinding(
                                     AnyJsBinding::JsIdentifierBinding(make::js_identifier_binding(
                                         make::ident("params"),
                                     )),
@@ -204,7 +204,7 @@ pub(crate) fn generate_workspace_bindings(mode: Mode) -> Result<()> {
             export = export.with_leading_trivia(trivia);
         }
         AnyJsModuleItem::JsExport(make::js_export(
-            make::js_decorator_list(std::iter::empty()),
+            make::js_decorator_list([]),
             export,
             AnyJsExportClause::AnyJsDeclarationClause(match decl {
                 AnyJsDeclaration::JsClassDeclaration(decl) => {
@@ -310,7 +310,7 @@ pub(crate) fn generate_workspace_bindings(mode: Mode) -> Result<()> {
     ));
 
     items.push(AnyJsModuleItem::JsExport(make::js_export(
-        make::js_decorator_list(std::iter::empty()),
+        make::js_decorator_list([]),
         make::token(T![export]),
         AnyJsExportClause::AnyJsDeclarationClause(AnyJsDeclarationClause::TsInterfaceDeclaration(
             make::ts_interface_declaration(
@@ -327,7 +327,7 @@ pub(crate) fn generate_workspace_bindings(mode: Mode) -> Result<()> {
     let member_separators = (0..member_declarations.len()).map(|_| make::token(T![,]));
 
     items.push(AnyJsModuleItem::JsExport(make::js_export(
-        make::js_decorator_list(std::iter::empty()),
+        make::js_decorator_list([]),
         make::token(T![export]),
         AnyJsExportClause::AnyJsDeclarationClause(AnyJsDeclarationClause::JsFunctionDeclaration(
             make::js_function_declaration(
@@ -341,7 +341,7 @@ pub(crate) fn generate_workspace_bindings(mode: Mode) -> Result<()> {
                         Some(AnyJsParameter::AnyJsFormalParameter(
                             AnyJsFormalParameter::JsFormalParameter(
                                 make::js_formal_parameter(
-                                    make::js_decorator_list(std::iter::empty()),
+                                    make::js_decorator_list([]),
                                     AnyJsBindingPattern::AnyJsBinding(
                                         AnyJsBinding::JsIdentifierBinding(
                                             make::js_identifier_binding(make::ident("transport")),